requirements.txt
torch
transformers
vllm
ctranslate2
onnxruntime-gpu
optimum
llama-cpp-python
huggingface_hub 